requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cssselect2==0.6.0
Django==4.0.4
django-ckeditor==6.4.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-js-asset==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
lxml==4.8.0
MarkupSafe==2.1.1
openapi-codec==1.3.2
oscrypto==1.3.0
Pillow==9.1.0
psycopg2-binary==2.9.3
pycparser==2.21
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
python-bidi==0.4.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.7
